Introduce tooling-spi to make HTML report generation extensible
