+++ cli/src/main/java/org/opentest4j/reporting/cli/ConvertCommand.java
@@ -16,7 +16,7 @@
 
 package org.opentest4j.reporting.cli;
 
-import org.opentest4j.reporting.tooling.converter.DefaultConverter;
+import org.opentest4j.reporting.tooling.core.converter.DefaultConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine.Command;
