+++ cli/src/main/java/org/opentest4j/reporting/cli/HtmlReportCommand.java
@@ -16,7 +16,7 @@
 
 package org.opentest4j.reporting.cli;
 
-import org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriter;
+import org.opentest4j.reporting.tooling.core.htmlreport.DefaultHtmlReportWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine.Command;
