+++ cli/src/main/java/org/opentest4j/reporting/cli/ValidateCommand.java
@@ -16,9 +16,9 @@
 
 package org.opentest4j.reporting.cli;
 
-import org.opentest4j.reporting.tooling.validator.DefaultValidator;
-import org.opentest4j.reporting.tooling.validator.ValidationResult;
-import org.opentest4j.reporting.tooling.validator.Validator;
+import org.opentest4j.reporting.tooling.core.validator.DefaultValidator;
+import org.opentest4j.reporting.tooling.core.validator.ValidationResult;
+import org.opentest4j.reporting.tooling.core.validator.Validator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine.Command;
@@ -30,8 +30,8 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.concurrent.Callable;
 
-import static org.opentest4j.reporting.tooling.validator.Severity.ERROR;
-import static org.opentest4j.reporting.tooling.validator.Severity.WARNING;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.ERROR;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.WARNING;
 
 /**
  * Validates XML report files of both formats.
