+++ gradle/libs.versions.toml
@@ -1,10 +1,15 @@
 [versions]
 eclipse = "4.32.0"
+immutables = "2.10.1"
+jackson = "2.18.0"
 
 [libraries]
 archunit = { module = "com.tngtech.archunit:archunit-junit5", version = "1.3.0" }
 assertj-core = { module = "org.assertj:assertj-core", version = "3.26.3" }
-jackson-dataformat-xml = { module = "com.fasterxml.jackson.dataformat:jackson-dataformat-xml", version = "2.18.0" }
+immutables-value = { module = "org.immutables:value", version.ref = "immutables" }
+immutables-valueAnnotations = { module = "org.immutables:value-annotations", version.ref = "immutables" }
+jackson-annotations = { module = "com.fasterxml.jackson.core:jackson-annotations", version.ref = "jackson" }
+jackson-dataformat-xml = { module = "com.fasterxml.jackson.dataformat:jackson-dataformat-xml", version.ref = "jackson" }
 jetbrains-annotations = { module = "org.jetbrains:annotations", version = "26.0.1" }
 joox = { module = "org.jooq:joox", version = "2.0.1" }
 jsonSanitizer = { module = "com.mikesamuel:json-sanitizer", version = "1.2.2" }
