+++ html-report/public/init.js
@@ -1,1 +1,1 @@
-globalThis.testExecutions = [{ "id": "1", "name": "events.xml", "durationMillis": 301, "sections": [{ "title": "Infrastructure", "blocks": [{ "type": "kvp", "content": { "Hostname": "coriol", "Username": "marc", "Operating system": "Linux", "CPU cores": "16" } }] }, { "title": "Java Virtual Machine", "blocks": [{ "type": "kvp", "content": { "Java version": "17.0.12", "File encoding": "UTF-8", "Max heap size": "512 MiB (536870912 bytes)" } }] }], "roots": ["2"], "children": { "2": ["3", "7", "11"], "3": ["4", "5", "6"], "7": ["8", "9", "10"], "11": ["12", "13", "14", "15"] }, "testNodes": [{ "id": "2", "name": "JUnit Jupiter", "durationMillis": 301, "status": "SUCCESSFUL" }, { "id": "3", "name": "DefaultConverterTests", "durationMillis": 128, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests" } }] }] }] }] }, { "id": "4", "name": "convertsInfrastructureSection()", "durationMillis": 70, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests", "methodName": "convertsInfrastructureSection" } }] }] }] }] }, { "id": "5", "name": "convertsStartedAndFinishedEvents()", "durationMillis": 24, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "ClasspathResource", "blocks": [{ "type": "kvp", "content": { "resourceName": "foo.xml", "line": "23", "column": "42" } }] }] }] }] }, { "id": "6", "name": "mergesReportEntries()", "durationMillis": 19, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests", "methodName": "mergesReportEntries" } }] }] }] }] }, { "id": "7", "name": "DefaultHtmlReportWriterTest", "durationMillis": 29, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest" } }] }] }] }] }, { "id": "8", "name": "aborted()", "durationMillis": 3, "status": "ABORTED", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "aborted" } }] }] }] }, { "title": "Result", "blocks": [{ "type": "sub", "content": [{ "title": "org.opentest4j.TestAbortedException", "blocks": [{ "type": "pre", "content": "org.opentest4j.TestAbortedException: message\n\tat org.junit.jupiter.api.Assumptions.abort(Assumptions.java:295)\n\tat org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest.aborted(DefaultHtmlReportWriterTest.java:66)\n\tat java.base/java.lang.reflect.Method.invoke(Method.java:569)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n" }] }] }] }] }, { "id": "9", "name": "Write sample HTML report", "durationMillis": 19, "status": "FAILED", "sections": [{ "title": "Tags", "blocks": [{ "type": "labels", "content": ["demo", "html-report", "tooling"] }] }, { "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "writeSampleHtmlReport", "methodParameterTypes": "org.junit.jupiter.api.TestReporter" } }] }] }] }, { "title": "Result", "blocks": [{ "type": "sub", "content": [{ "title": "org.xml.sax.SAXParseException", "blocks": [{ "type": "pre", "content": "org.xml.sax.SAXParseException; systemId: file:/home/marc/Repositories/open-test-reporting/tooling/build/test-results/test/junit-platform-events-1674224502021996829.xml; lineNumber: 24; columnNumber: 391; XML document structures must start and end within the same entity.\n\tat java.xml/com.sun.org.apache.xerces.internal.parsers.DOMParser.parse(DOMParser.java:262)\n\tat java.xml/com.sun.org.apache.xerces.internal.jaxp.DocumentBuilderImpl.parse(DocumentBuilderImpl.java:342)\n\tat java.xml/javax.xml.parsers.DocumentBuilder.parse(DocumentBuilder.java:206)\n\tat org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriter.parseDom(DefaultHtmlReportWriter.java:399)\n\tat org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriter.writeHtmlReport(DefaultHtmlReportWriter.java:51)\n\tat org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest.writeSampleHtmlReport(DefaultHtmlReportWriterTest.java:48)\n\tat java.base/java.lang.reflect.Method.invoke(Method.java:569)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n" }] }] }] }, { "title": "Attachments", "blocks": [{ "type": "sub", "content": [{ "title": "Data", "blocks": [{ "type": "kvp", "content": { "time": "2024-11-05T15:01:42.474827283", "value": "Start" } }] }, { "title": "Data", "blocks": [{ "type": "kvp", "content": { "time": "2024-11-05T15:01:42.477377450", "targetFile": "build/test-results/open-test-report.html", "eventsXml": "build/test-results/test/junit-platform-events-1674224502021996829.xml", "hierarchyXml": "build/test-results/test/hierarchy.xml" } }] }] }] }] }, { "id": "10", "name": "skipped()", "durationMillis": 0, "status": "SKIPPED", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "skipped" } }] }] }] }, { "title": "Result", "blocks": [{ "type": "sub", "content": [{ "title": "Reason", "blocks": [{ "type": "p", "content": "Disabled on JRE version: 17.0.12" }] }] }] }] }, { "id": "11", "name": "DefaultValidatorTests", "durationMillis": 127, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests" } }] }] }] }] }, { "id": "12", "name": "validatesEventsXmlFile()", "durationMillis": 34, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "validatesEventsXmlFile" } }] }] }] }] }, { "id": "13", "name": "generatesWarningsAndErrors()", "durationMillis": 78, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "generatesWarningsAndErrors" } }] }] }] }] }, { "id": "14", "name": "resolvesThirdPartySchemasUsingCatalog()", "durationMillis": 4, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "resolvesThirdPartySchemasUsingCatalog" } }] }] }] }] }, { "id": "15", "name": "validatesHierarchicalXmlFile()", "durationMillis": 5, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "validatesHierarchicalXmlFile" } }] }] }] }] }] }, { "id": "16", "name": "hierarchy.xml", "durationMillis": 472, "sections": [{ "title": "Infrastructure", "blocks": [{ "type": "kvp", "content": { "Hostname": "coriol", "Username": "marc", "Operating system": "Linux", "CPU cores": "16" } }] }, { "title": "Java Virtual Machine", "blocks": [{ "type": "kvp", "content": { "Java version": "17.0.12", "File encoding": "UTF-8", "Max heap size": "512 MiB (536870912 bytes)" } }] }], "roots": ["17"], "children": { "17": ["18", "22", "26"], "18": ["19", "20", "21"], "22": ["23", "24", "25"], "26": ["27", "28", "29", "30"] }, "testNodes": [{ "id": "17", "name": "JUnit Jupiter", "durationMillis": 472, "status": "SUCCESSFUL" }, { "id": "18", "name": "DefaultConverterTests", "durationMillis": 138, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests" } }] }] }] }] }, { "id": "19", "name": "convertsInfrastructureSection()", "durationMillis": 81, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests", "methodName": "convertsInfrastructureSection" } }] }] }] }] }, { "id": "20", "name": "convertsStartedAndFinishedEvents()", "durationMillis": 22, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests", "methodName": "convertsStartedAndFinishedEvents" } }] }] }] }] }, { "id": "21", "name": "mergesReportEntries()", "durationMillis": 18, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.converter.DefaultConverterTests", "methodName": "mergesReportEntries" } }] }] }] }] }, { "id": "22", "name": "DefaultHtmlReportWriterTest", "durationMillis": 164, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest" } }] }] }] }] }, { "id": "23", "name": "aborted()", "durationMillis": 4, "status": "ABORTED", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "aborted" } }] }] }] }, { "title": "Result", "blocks": [{ "type": "sub", "content": [{ "title": "org.opentest4j.TestAbortedException", "blocks": [{ "type": "pre", "content": "org.opentest4j.TestAbortedException: message\n\tat org.junit.jupiter.api.Assumptions.abort(Assumptions.java:295)\n\tat org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest.aborted(DefaultHtmlReportWriterTest.java:60)\n\tat java.base/java.lang.reflect.Method.invoke(Method.java:569)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n\tat java.base/java.util.ArrayList.forEach(ArrayList.java:1511)\n" }] }] }] }] }, { "id": "24", "name": "Write sample HTML report", "durationMillis": 151, "status": "SUCCESSFUL", "sections": [{ "title": "Tags", "blocks": [{ "type": "labels", "content": ["demo", "html-report", "tooling"] }] }, { "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "writeSampleHtmlReport", "methodParameterTypes": "org.junit.jupiter.api.TestReporter" } }] }] }] }, { "title": "Attachments", "blocks": [{ "type": "sub", "content": [{ "title": "Data", "blocks": [{ "type": "kvp", "content": { "time": "2024-11-05T17:02:08.025302623", "value": "Start" } }] }, { "title": "Data", "blocks": [{ "type": "kvp", "content": { "time": "2024-11-05T17:02:08.027741131", "targetFile": "build/test-results/test/open-test-report.html", "hierarchyXml": "build/test-results/test/hierarchy.xml", "eventsXml": "build/test-results/test/events.xml" } }] }] }] }] }, { "id": "25", "name": "skipped()", "durationMillis": 1, "status": "SKIPPED", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.htmlreport.DefaultHtmlReportWriterTest", "methodName": "skipped" } }] }] }] }, { "title": "Result", "blocks": [{ "type": "sub", "content": [{ "title": "Reason", "blocks": [{ "type": "p", "content": "Disabled on JRE version: 17.0.12" }] }] }] }] }, { "id": "26", "name": "DefaultValidatorTests", "durationMillis": 148, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests" } }] }] }] }] }, { "id": "27", "name": "validatesEventsXmlFile()", "durationMillis": 44, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "validatesEventsXmlFile" } }] }] }] }] }, { "id": "28", "name": "generatesWarningsAndErrors()", "durationMillis": 87, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "generatesWarningsAndErrors" } }] }] }] }] }, { "id": "29", "name": "resolvesThirdPartySchemasUsingCatalog()", "durationMillis": 6, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "resolvesThirdPartySchemasUsingCatalog" } }] }] }] }] }, { "id": "30", "name": "validatesHierarchicalXmlFile()", "durationMillis": 6, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.validator.DefaultValidatorTests", "methodName": "validatesHierarchicalXmlFile" } }] }] }] }] }] }];
+globalThis.testExecutions = [{ "id": "1", "name": "junit-platform-events-835360283716872309.xml", "durationMillis": 279, "sections": [{ "title": "Infrastructure", "blocks": [{ "type": "kvp", "content": { "Hostname": "coriol", "Username": "marc", "Operating system": "Linux", "CPU cores": "16" } }] }, { "title": "Java Virtual Machine", "blocks": [{ "type": "kvp", "content": { "Java version": "17.0.12", "File encoding": "UTF-8", "Max heap size": "512 MiB (536870912 bytes)" } }] }], "roots": ["2"], "children": { "2": ["3", "7"], "3": ["4", "5", "6"], "7": ["8", "9", "10", "11"] }, "testNodes": [{ "id": "2", "name": "JUnit Jupiter", "durationMillis": 279, "status": "SUCCESSFUL" }, { "id": "3", "name": "DefaultConverterTests", "durationMillis": 129, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.converter.DefaultConverterTests" } }] }] }] }] }, { "id": "4", "name": "convertsInfrastructureSection()", "durationMillis": 76, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.converter.DefaultConverterTests", "methodName": "convertsInfrastructureSection" } }] }] }] }] }, { "id": "5", "name": "convertsStartedAndFinishedEvents()", "durationMillis": 21, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.converter.DefaultConverterTests", "methodName": "convertsStartedAndFinishedEvents" } }] }] }] }] }, { "id": "6", "name": "mergesReportEntries()", "durationMillis": 18, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.converter.DefaultConverterTests", "methodName": "mergesReportEntries" } }] }] }] }] }, { "id": "7", "name": "DefaultValidatorTests", "durationMillis": 132, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Class", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.validator.DefaultValidatorTests" } }] }] }] }] }, { "id": "8", "name": "validatesEventsXmlFile()", "durationMillis": 35, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.validator.DefaultValidatorTests", "methodName": "validatesEventsXmlFile" } }] }] }] }] }, { "id": "9", "name": "generatesWarningsAndErrors()", "durationMillis": 79, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.validator.DefaultValidatorTests", "methodName": "generatesWarningsAndErrors" } }] }] }] }] }, { "id": "10", "name": "resolvesThirdPartySchemasUsingCatalog()", "durationMillis": 6, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.validator.DefaultValidatorTests", "methodName": "resolvesThirdPartySchemasUsingCatalog" } }] }] }] }] }, { "id": "11", "name": "validatesHierarchicalXmlFile()", "durationMillis": 7, "status": "SUCCESSFUL", "sections": [{ "title": "Sources", "blocks": [{ "type": "sub", "content": [{ "title": "Method", "blocks": [{ "type": "kvp", "content": { "className": "org.opentest4j.reporting.tooling.core.validator.DefaultValidatorTests", "methodName": "validatesHierarchicalXmlFile" } }] }] }] }] }] }];
