+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/converter/Converter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.converter;
+package org.opentest4j.reporting.tooling.core.converter;
 
 import java.nio.file.Path;
