+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/converter/DefaultConverter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.converter;
+package org.opentest4j.reporting.tooling.core.converter;
 
 import org.opentest4j.reporting.events.core.Infrastructure;
 import org.opentest4j.reporting.events.root.Event;
@@ -23,7 +23,7 @@ import org.opentest4j.reporting.events.root.Reported;
 import org.opentest4j.reporting.events.root.Started;
 import org.opentest4j.reporting.schema.Namespace;
 import org.opentest4j.reporting.schema.QualifiedName;
-import org.opentest4j.reporting.tooling.util.DomUtils;
+import org.opentest4j.reporting.tooling.core.util.DomUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -51,7 +51,7 @@ import static java.util.stream.Collectors.toUnmodifiableSet;
 import static javax.xml.XMLConstants.FEATURE_SECURE_PROCESSING;
 import static javax.xml.XMLConstants.XMLNS_ATTRIBUTE;
 import static javax.xml.XMLConstants.XMLNS_ATTRIBUTE_NS_URI;
-import static org.opentest4j.reporting.tooling.util.DomUtils.stream;
+import static org.opentest4j.reporting.tooling.core.util.DomUtils.stream;
 
 /**
  * Default implementation of {@link Converter}.
