+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/htmlreport/CoreContributor.java
@@ -1,0 +1,164 @@
+package org.opentest4j.reporting.tooling.core.htmlreport;
+
+import org.joox.Match;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Contributor;
+import org.opentest4j.reporting.tooling.spi.htmlreport.KeyValuePairs;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Labels;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Paragraph;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Section;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Subsections;
+import org.w3c.dom.Element;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Optional;
+
+import static org.joox.JOOX.$;
+import static org.opentest4j.reporting.tooling.core.util.DomUtils.stream;
+
+public class CoreContributor implements Contributor {
+
+    @Override
+    public List<Section> contributeSectionsForExecution(Element executionElement) {
+        var sections = new ArrayList<Section>();
+        createInfrastructureSection(executionElement).ifPresent(sections::add);
+        return sections;
+    }
+
+    @Override
+    public List<Section> contributeSectionsForTestNode(Element testNodeElement) {
+        var sections = new ArrayList<Section>();
+        createTagsSection(testNodeElement).ifPresent(sections::add);
+        createSourcesSection(testNodeElement).ifPresent(sections::add);
+        createReasonSection(testNodeElement).ifPresent(sections::add);
+        createAttachmentsSection(testNodeElement).ifPresent(sections::add);
+        return sections;
+    }
+
+    private static Optional<Section> createInfrastructureSection(Element element) {
+        var infrastructure = $(element).child("infrastructure");
+
+        var table = new LinkedHashMap<String, String>();
+        addToTable(infrastructure, "hostName", "Hostname", table);
+        addToTable(infrastructure, "userName", "Username", table);
+        addToTable(infrastructure, "operatingSystem", "Operating system", table);
+        addToTable(infrastructure, "cpuCores", "CPU cores", table);
+
+        if (table.isEmpty()) {
+            return Optional.empty();
+        }
+
+        return Optional.of(Section.builder()
+                .title("Infrastructure")
+                .addBlocks(KeyValuePairs.builder().content(table).build())
+                .build());
+    }
+
+    private static Optional<Section> createTagsSection(Element element) {
+        var tags = $(element)
+                .child("metadata")
+                .child("tags")
+                .children("tag")
+                .map(e -> e.element().getTextContent());
+        if (tags.isEmpty()) {
+            return Optional.empty();
+        }
+        var section = Section.builder()
+                .title("Tags")
+                .addBlocks(Labels.builder().content(tags).build())
+                .build();
+        return Optional.of(section);
+    }
+
+    private static Optional<Section> createSourcesSection(Element element) {
+        var children = $(element).children("sources").children().get();
+        if (children.isEmpty()) {
+            return Optional.empty();
+        }
+        var subsections = Subsections.builder();
+        children.stream()
+                .map(child -> {
+                    var type = child.getLocalName();
+                    if (child.getLocalName().endsWith("Source")) {
+                        type = child.getLocalName().substring(0, child.getLocalName().length() - "Source".length());
+                    }
+                    var subsection = Section.builder().title(capitalize(type));
+                    var attributes = KeyValuePairs.builder();
+                    stream(child.getAttributes()).forEach(it -> attributes.putContent(it.getNodeName(), it.getNodeValue()));
+                    var filePosition = $(child).child("filePosition").first();
+                    if (filePosition.isNotEmpty()) {
+                        attributes.putContent("line", filePosition.attr("line"));
+                        var column = filePosition.attr("column");
+                        if (column != null) {
+                            attributes.putContent("column", column);
+                        }
+                    }
+                    subsection.addBlocks(attributes.build());
+                    return subsection.build();
+                })
+                .forEach(subsections::addContent);
+
+        return Optional.of(Section.builder()
+                .title("Sources")
+                .addBlocks(subsections.build())
+                .build());
+    }
+
+    private static Optional<Section> createReasonSection(Element element) {
+        var reason = $(element).child("result").child("reason").text();
+        if (reason == null) {
+            return Optional.empty();
+        }
+        return Optional.of(Section.builder()
+                .title("Reason")
+                .addBlocks(Paragraph.builder().content(reason).build())
+                .build());
+    }
+
+    private static Optional<Section> createAttachmentsSection(Element element) {
+        var children = $(element).children("attachments").children().get();
+        if (children.isEmpty()) {
+            return Optional.empty();
+        }
+
+        var subsections = Subsections.builder();
+        children.stream()
+                .map(child -> {
+                    var attributes = KeyValuePairs.builder();
+                    stream(child.getAttributes()).forEach(it -> attributes.putContent(it.getNodeName(), it.getNodeValue()));
+                    if ("data".equals(child.getLocalName())) {
+                        $(child).children("entry").forEach(entry -> {
+                            var key = entry.getAttribute("key");
+                            var value = entry.getTextContent();
+                            attributes.putContent(key, value);
+                        });
+                    }
+                    var type = child.getLocalName();
+                    return Section.builder()
+                            .title(capitalize(type))
+                            .addBlocks(attributes.build())
+                            .build();
+                })
+                .forEach(subsections::addContent);
+
+        return Optional.of(Section.builder()
+                .title("Attachments")
+                .addBlocks(subsections.build())
+                .build());
+    }
+
+    static void addToTable(Match march, String elementName, String label, Map<String, String> table) {
+        var value = march.child(elementName).text();
+        if (value != null) {
+            table.put(label, value);
+        }
+    }
+
+    private static String capitalize(String value) {
+        return value.substring(0, 1).toUpperCase(Locale.ROOT) + value.substring(1);
+    }
+
+}
