+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/htmlreport/DefaultHtmlReportWriter.java
@@ -1,0 +1,199 @@
+package org.opentest4j.reporting.tooling.core.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.json.JsonSanitizer;
+import org.opentest4j.reporting.events.core.Result;
+import org.opentest4j.reporting.events.root.Events;
+import org.opentest4j.reporting.schema.Namespace;
+import org.opentest4j.reporting.tooling.core.converter.DefaultConverter;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Contributor;
+import org.opentest4j.reporting.tooling.spi.htmlreport.Section;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.xml.sax.SAXException;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.Duration;
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Deque;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.ServiceLoader;
+import java.util.function.BiFunction;
+
+import static java.util.Comparator.comparing;
+import static java.util.Objects.requireNonNull;
+import static org.opentest4j.reporting.tooling.core.util.DomUtils.matches;
+import static org.opentest4j.reporting.tooling.core.util.DomUtils.stream;
+
+public class DefaultHtmlReportWriter implements HtmlReportWriter {
+
+    private final ServiceLoader<Contributor> contributors;
+
+    public DefaultHtmlReportWriter() {
+        contributors = ServiceLoader.load(Contributor.class);
+    }
+
+    @Override
+    public void writeHtmlReport(List<Path> xmlFiles, Path htmlFile) throws Exception {
+
+        var idGenerator = new IdGenerator();
+        var executions = new ArrayList<Execution>();
+
+        for (Path xmlFile : xmlFiles) {
+            var rootElement = parseDom(xmlFile);
+            var name = String.format(xmlFile.getFileName().toString());
+            executions.add(extractData(idGenerator, rootElement, name));
+        }
+
+        var javaScript = toJavascript(executions);
+
+        var content = readTemplate().replace("<script src=\"./init.js\"></script>", "<script>\n    " + javaScript + "\n  </script>");
+        Files.writeString(htmlFile, content);
+    }
+
+    private static String toJavascript(ArrayList<Execution> executions) throws JsonProcessingException {
+        var jsonMapper = new ObjectMapper();
+        jsonMapper.setSerializationInclusion(JsonInclude.Include.NON_EMPTY);
+
+        var json = JsonSanitizer.sanitize(jsonMapper.writeValueAsString(executions));
+        return "globalThis.testExecutions = " + json + ";";
+    }
+
+    private Execution extractData(IdGenerator idGenerator, Element rootElement, String name) {
+        var execution = new Execution(idGenerator.next(), name);
+        addSections(rootElement, execution);
+
+        for (Node child = rootElement.getFirstChild(); child != null; child = child.getNextSibling()) {
+            if (child instanceof Element && Namespace.REPORTING_HIERARCHY.getUri().equals(child.getNamespaceURI())
+                    && "root".equals(child.getLocalName())) {
+                var root = visitNode((Element) child, idGenerator, new ArrayDeque<>(), execution);
+                execution.durationMillis += root.durationMillis;
+                execution.roots.add(root.id);
+            }
+        }
+        return execution;
+    }
+
+    static class Execution {
+
+        public final String id;
+        public final String name;
+        public long durationMillis;
+
+        public final List<Section> sections = new ArrayList<>();
+
+        public final List<String> roots = new ArrayList<>();
+        public final Map<String, List<String>> children = new LinkedHashMap<>();
+        public final List<TestNode> testNodes = new ArrayList<>();
+
+        Execution(String id, String name) {
+            this.id = id;
+            this.name = name;
+        }
+    }
+
+    static class TestNode {
+
+        public final String id;
+        public final String name;
+        public final long durationMillis;
+
+        public final String status;
+
+        public final List<Section> sections = new ArrayList<>();
+
+        TestNode(String id, String name, long durationMillis, String status) {
+            this.id = id;
+            this.name = name;
+            this.durationMillis = durationMillis;
+            this.status = status;
+        }
+    }
+
+    private TestNode visitNode(Element node, IdGenerator idGenerator, Deque<String> parentIds, Execution execution) {
+        String currentId = idGenerator.next();
+
+        String status = stream(node.getChildNodes()) //
+                .filter(it -> it instanceof Element && matches(Result.ELEMENT, (Element) it))
+                .findAny()
+                .map(it -> ((Element) it).getAttribute(Result.STATUS.getSimpleName()))
+                .orElse(null);
+
+        var duration = Duration.parse(node.getAttribute("duration"));
+        var testNode = new TestNode(currentId, node.getAttribute("name"), duration.toMillis(), status);
+        execution.testNodes.add(testNode);
+
+        addSections(node, testNode);
+
+        if (!parentIds.isEmpty()) {
+            String parentId = parentIds.peek();
+            execution.children.computeIfAbsent(parentId, __ -> new ArrayList<>());
+            execution.children.get(parentId).add(currentId);
+        }
+        parentIds.push(currentId);
+        for (Node child = node.getFirstChild(); child != null; child = child.getNextSibling()) {
+            if (child instanceof Element && Namespace.REPORTING_HIERARCHY.getUri().equals(child.getNamespaceURI())
+                    && "child".equals(child.getLocalName())) {
+                visitNode((Element) child, idGenerator, parentIds, execution);
+            }
+        }
+        parentIds.pop();
+
+        return testNode;
+    }
+
+    private void addSections(Element element, Execution execution) {
+        contributeSections(element, Contributor::contributeSectionsForExecution, execution.sections);
+    }
+
+    private void addSections(Element element, TestNode node) {
+        contributeSections(element, Contributor::contributeSectionsForTestNode, node.sections);
+    }
+
+    private void contributeSections(Element element, BiFunction<Contributor, Element, List<Section>> call, List<Section> sections) {
+        contributors.stream()
+                .flatMap(it -> call.apply(it.get(), element).stream())
+                .sorted(comparing(Section::getOrder))
+                .forEach(sections::add);
+    }
+
+    static class IdGenerator {
+        private int nextId = 1;
+
+        String next() {
+            return Integer.toString(nextId++);
+        }
+    }
+
+    private static Element parseDom(Path xmlFile) throws ParserConfigurationException, SAXException, IOException {
+        DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+        factory.setNamespaceAware(true);
+        DocumentBuilder builder = factory.newDocumentBuilder();
+
+        Document sourceDocument = builder.parse(xmlFile.toFile());
+        Element element = sourceDocument.getDocumentElement();
+
+        if (matches(Events.ELEMENT, element)) {
+            return new DefaultConverter().convert(element).getDocumentElement();
+        }
+
+        return element;
+    }
+
+    private String readTemplate() throws IOException {
+        try (var template = getClass().getResourceAsStream("template.html")) {
+            return new String(requireNonNull(template).readAllBytes());
+        }
+    }
+}
