+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/htmlreport/HtmlReportWriter.java
@@ -1,4 +1,4 @@
-package org.opentest4j.reporting.tooling.htmlreport;
+package org.opentest4j.reporting.tooling.core.htmlreport;
 
 import java.nio.file.Path;
 import java.util.List;
