+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/util/DomUtils.java
@@ -1,4 +1,4 @@
-package org.opentest4j.reporting.tooling.util;
+package org.opentest4j.reporting.tooling.core.util;
 
 import org.opentest4j.reporting.schema.QualifiedName;
 import org.w3c.dom.Element;
