+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/validator/DefaultValidator.java
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.validator;
+package org.opentest4j.reporting.tooling.core.validator;
 
 import org.opentest4j.reporting.schema.Namespace;
-import org.opentest4j.reporting.tooling.converter.Converter;
 import org.w3c.dom.ls.LSInput;
 import org.w3c.dom.ls.LSResourceResolver;
 import org.xml.sax.ErrorHandler;
@@ -44,8 +43,8 @@ import java.util.Map;
 import static java.util.Objects.requireNonNull;
 import static javax.xml.XMLConstants.W3C_XML_SCHEMA_NS_URI;
 import static javax.xml.catalog.CatalogManager.catalogResolver;
-import static org.opentest4j.reporting.tooling.validator.Severity.ERROR;
-import static org.opentest4j.reporting.tooling.validator.Severity.WARNING;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.ERROR;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.WARNING;
 
 /**
  * Default implementation of {@link Validator}.
