+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/validator/Severity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.validator;
+package org.opentest4j.reporting.tooling.core.validator;
 
 /**
  * Severity of a {@linkplain ValidationMessage validation message}.
