+++ tooling-core/src/main/java/org/opentest4j/reporting/tooling/core/validator/Validator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.validator;
+package org.opentest4j.reporting.tooling.core.validator;
 
 import java.nio.file.Path;
