+++ tooling-core/src/main/resources/META-INF/services/org.opentest4j.reporting.tooling.spi.htmlreport.Contributor
@@ -1,0 +1,2 @@
+org.opentest4j.reporting.tooling.core.htmlreport.CoreContributor
+org.opentest4j.reporting.tooling.core.htmlreport.JavaContributor
