+++ tooling-core/src/test/java/org/opentest4j/reporting/tooling/core/converter/DefaultConverterTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.converter;
+package org.opentest4j.reporting.tooling.core.converter;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
