+++ tooling-core/src/test/java/org/opentest4j/reporting/tooling/core/validator/DefaultValidatorTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.opentest4j.reporting.tooling.validator;
+package org.opentest4j.reporting.tooling.core.validator;
 
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.Test;
@@ -26,8 +26,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.opentest4j.reporting.tooling.validator.Severity.ERROR;
-import static org.opentest4j.reporting.tooling.validator.Severity.WARNING;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.ERROR;
+import static org.opentest4j.reporting.tooling.core.validator.Severity.WARNING;
 
 class DefaultValidatorTests {
