+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Block.java
@@ -1,0 +1,12 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+import static com.fasterxml.jackson.annotation.JsonTypeInfo.Id.SIMPLE_NAME;
+
+@JsonTypeInfo(use = SIMPLE_NAME, property = "type")
+public interface Block<T> {
+
+    T getContent();
+
+}
