+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Contributor.java
@@ -1,0 +1,29 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import org.w3c.dom.Element;
+
+import java.util.List;
+
+import static java.util.Collections.emptyList;
+
+/**
+ * Implementations of this interface can contribute additional sections to the
+ * HTML report when registered via the Java service loader mechanism.
+ */
+public interface Contributor {
+
+    /**
+     * Contribute additional sections for the given execution element.
+     */
+    default List<Section> contributeSectionsForExecution(Element executionElement) {
+        return emptyList();
+    }
+
+    /**
+     * Contribute additional sections for the given test node element.
+     */
+    default List<Section> contributeSectionsForTestNode(Element testNodeElement) {
+        return emptyList();
+    }
+
+}
