+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/KeyValuePairs.java
@@ -1,0 +1,19 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonTypeName;
+import org.immutables.value.Value.Immutable;
+
+import java.util.Map;
+
+@Immutable
+@JsonTypeName("kvp")
+public interface KeyValuePairs extends Block<Map<String, String>> {
+
+    class Builder extends ImmutableKeyValuePairs.Builder {
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+}
