+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Labels.java
@@ -1,0 +1,19 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonTypeName;
+import org.immutables.value.Value.Immutable;
+
+import java.util.List;
+
+@Immutable
+@JsonTypeName("labels")
+public interface Labels extends Block<List<String>> {
+
+    class Builder extends ImmutableLabels.Builder {
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+}
