+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Paragraph.java
@@ -1,0 +1,17 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonTypeName;
+import org.immutables.value.Value.Immutable;
+
+@Immutable
+@JsonTypeName("p")
+public interface Paragraph extends Block<String> {
+
+    class Builder extends ImmutableParagraph.Builder {
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+}
