+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Section.java
@@ -1,0 +1,29 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.immutables.value.Value.Default;
+import org.immutables.value.Value.Immutable;
+
+import java.util.List;
+
+@Immutable
+public interface Section {
+
+    class Builder extends ImmutableSection.Builder {
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+    String getTitle();
+
+    List<Block<?>> getBlocks();
+
+    @JsonIgnore
+    @Default
+    default int getOrder() {
+        return 0;
+    }
+
+}
