+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/Subsections.java
@@ -1,0 +1,19 @@
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import com.fasterxml.jackson.annotation.JsonTypeName;
+import org.immutables.value.Value.Immutable;
+
+import java.util.List;
+
+@Immutable
+@JsonTypeName("sub")
+public interface Subsections extends Block<List<Section>> {
+
+    class Builder extends ImmutableSubsections.Builder {
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+}
