+++ tooling-spi/src/main/java/org/opentest4j/reporting/tooling/spi/htmlreport/package-info.java
@@ -1,0 +1,13 @@
+@Style(
+        visibility = ImplementationVisibility.PACKAGE,
+        builderVisibility = BuilderVisibility.PACKAGE,
+        jdkOnly = true,
+        overshadowImplementation = true,
+        from = "",
+        with = ""
+)
+package org.opentest4j.reporting.tooling.spi.htmlreport;
+
+import org.immutables.value.Value.Style;
+import org.immutables.value.Value.Style.BuilderVisibility;
+import org.immutables.value.Value.Style.ImplementationVisibility;
